feat: colorize shell command output
This commit introduces a new feature that automatically colorizes
the output of shell commands in Emacs.  It achieves this by advising
the `shell-command-on-region` function to run `cat-colorize-buffer`
on the `*Shell Command Output*` buffer after the command has
finished executing.

Additionally, the `ansi-color-for-comint-mode-on` hook is added to
`shell-mode` to ensure that ANSI colors are properly displayed in
interactive shell sessions.