feat: extract task related functions to task.el
This commit introduces a new file, `elisp/task.el`, to house
task-related functions, specifically those interacting with JIRA and
Git.  It moves the functions `task--generate-branch-name` and
`task-create-branch-with-key-and-text` from `cats/+org-jira.el` to
`elisp/task.el`.  It also introduces a function
`task-jira-select-issue` to allow users to select a JIRA issue using
JQL.  Lastly, it updates `+org-jira-start-dev-work` to use the new
`task-create-branch-with-key-and-text` function.